#165273577 Like or Dislike An Article

+++ authors/apps/articles/views.py
@@ -25,7 +25,7 @@ from rest_framework.permissions import (
 
 from .models import Article
 from .serializers import ArticleSerializer
-from .extra_methods import create_slug
+from .extra_methods import create_slug, like_grand_count
 from .mixins import CustomPaginationMixin
 from .extra_methods import create_slug
 from authors.apps.article_tag.views import ArticleTagViewSet
@@ -36,8 +36,7 @@ from authors.apps.profiles.serializers import ProfileSerializer
 from authors.apps.profiles.models import Profile
 from authors.apps.authentication.models import User
 
-
-class Articles(APIView, CustomPaginationMixin):
+class ListCreateArticlesView(APIView, CustomPaginationMixin):
     """
     This deals with;
     1. Getting all articles from the db
@@ -55,10 +54,16 @@ class Articles(APIView, CustomPaginationMixin):
         if page is not None:
             serializer = self.serializer_class(page, many=True)
             articles_count = len(serializer.data)
+            all_articles = serializer.data
 
+            articles_list = []
+
+            for article in all_articles:
+                article['likeCount'] = [like_grand_count(article)]
+                articles_list.append(article)
             return self.get_paginated_response(
                 {
-                    "articles": serializer.data,
+                    "articles": articles_list,
                     "articlesCount": articles_count,
                     "status": status.HTTP_200_OK,
                 }
@@ -74,16 +79,18 @@ class Articles(APIView, CustomPaginationMixin):
         """create tag if it doesn't exist"""
         tag_list = ArticleTagViewSet.create_tag_if_not_exist(ArticleTagViewSet, data.get('tagList'))
         data["tagList"] = tag_list
+
         serializer = ArticleSerializer(data=data)
         if serializer.is_valid():
             serializer.save()
-            article = serializer.data
+            article = dict(serializer.data)
+            article['likeCount'] = [like_grand_count(article)]
             return Response({"article": article}, status.HTTP_201_CREATED)
 
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-class OneArticle(APIView):
+class UpdateDeleteArticleView(APIView):
     """
     This deals with;
     1. Getting a single article from the db
@@ -94,46 +101,65 @@ class OneArticle(APIView):
     # Route protection
     permission_classes = (IsAuthenticated,)
 
-    def get(cls, request, slug):
-        article = get_object_or_404(Article, slug=slug, delete_status=False)
+    def get(cls, request, article_id):
+        article = get_object_or_404(Article, id=article_id, delete_status=False)
         serializer = ArticleSerializer(article, many=False)
         an_article = dict(serializer.data)
+        an_article['likeCount'] = [like_grand_count(an_article)]
         return Response({"article": an_article}, status=status.HTTP_200_OK)
-
-    def put(cls, request, slug):
-        article = get_object_or_404(Article, slug=slug, delete_status=False)
+    
+    def put(cls, request, article_id):
+        article = get_object_or_404(Article, id=article_id, delete_status=False)
         serializer = ArticleSerializer(article, many=False)
         json_data = request.data
-        json_data = request.data["article"]
-        data = dict(serializer.data)
-        data["updatedAt"] = timezone.now()
-        data["slug"] = create_slug(Article, data["title"])
-        data["title"] = json_data.get("title")
-        data["body"] = json_data.get("body")
-        data["description"] = json_data.get("description")
+        current_user = request.user
+        article_data = dict(serializer.data)
+        author = article_data["author"]
 
-        """create tag if it doesn't exist"""
-        tag_list = ArticleTagViewSet.create_tag_if_not_exist(ArticleTagViewSet, json_data.get('tagList'))
-        data["tagList"] = tag_list
+        if current_user.username == author:
+            json_data = request.data["article"]
+            article_data["updatedAt"] = timezone.now()
+            article_data["slug"] = create_slug(Article, article_data['title'])
+            article_data["title"] = json_data.get("title")
+            article_data["body"] = json_data.get("body")
+            article_data["description"] = json_data.get("description")
 
-        serializer = ArticleSerializer(article, data)
-        if serializer.is_valid():
-            serializer.save()
-            article = serializer.data
-            return Response({"article": article}, status=status.HTTP_200_OK)
-        return Response(
-                serializer.errors, status=status.HTTP_400_BAD_REQUEST
+            """create tag if it doesn't exist"""
+            tag_list = ArticleTagViewSet.create_tag_if_not_exist(ArticleTagViewSet, json_data.get('tagList'))
+            article_data["tagList"] = tag_list
+           
+            serializer = ArticleSerializer(article, article_data)
+
+            if serializer.is_valid():
+                serializer.save()
+                article = dict(serializer.data)
+                article['likeCount'] = [like_grand_count(article)]
+                return Response({"article": article}, status=status.HTTP_200_OK)
+            return Response(
+                    serializer.errors, status=status.HTTP_400_BAD_REQUEST
             )
 
-    def delete(cls, request, slug):
-        article = get_object_or_404(Article, slug=slug, delete_status=False)
-        article.delete()
         return Response(
-            {"messege": "article deleted successfully"},
-            status=status.HTTP_200_OK,
+                {"error":"you cannot edit an article created by another user"},
+                status=status.HTTP_401_UNAUTHORIZED
         )
 
+    def delete(cls, request, article_id):
+        article = get_object_or_404(Article, id=article_id, delete_status=False)
+        serializer = ArticleSerializer(article, many=False)
+        current_user = request.user        
+        article_data = dict(serializer.data)
+        author = article_data["author"]
 
+        if current_user.username == author:
+            article = get_object_or_404(Article, id=article_id, delete_status=False)
+            article.delete()
+            return Response({"messege": "article deleted successfully"}, status=status.HTTP_200_OK)
+        return Response(
+                {"error": "you cannot delete an article created by another user"},
+                status=status.HTTP_401_UNAUTHORIZED
+        )
+    
 class FavoriteArticleCreate(generics.ListCreateAPIView):
     """If the user feels satisfied with the article, he can favourite it """
 
@@ -204,6 +230,4 @@ class UnFavoriteArticleDestroy(generics.DestroyAPIView):
         return Response(
             {
                 "message": "You have successfully unfavorited this article."
-            },
-            status.HTTP_200_OK
-        )
+            }, status.HTTP_200_OK)
